fix: printRankList() 메서드를 Game 클래스에서 GamePrint 클래스로 이동
